pyproject.toml
[project]
name = "recipe-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.22.0",
    "ipykernel>=7.1.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "numpy==2.4.1",
    "pandas==2.3.3",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
